[MPT] Test partitioning in order not to ignore long running tests
