+++ build.sbt
@@ -1,14 +1,23 @@
-name := "etc-client"
 
-version := "0.1"
+val commonSettings = Seq(
+  name := "etc-client",
+  version := "0.1",
+  scalaVersion := "2.11.8"
+)
 
-scalaVersion := "2.11.8"
-
-libraryDependencies ++= Seq(
+val dep = Seq(
   "com.typesafe.akka" %% "akka-actor" % "2.4.16",
   "org.consensusresearch" %% "scrypto" % "1.2.0-RC3",
   "com.madgag.spongycastle" % "core" % "1.54.0.0",
-  "org.scalatest" %% "scalatest" % "3.0.1" % "test",
-  "org.scalacheck" %% "scalacheck" % "1.13.4" % "test",
+  "org.scalatest" %% "scalatest" % "3.0.1" % "it,test",
+  "org.scalacheck" %% "scalacheck" % "1.13.4" % "it,test",
   "org.scorexfoundation" %% "iodb" % "0.1.1"
 )
+
+val Integration = config("it") extend Test
+
+val root = project.in(file("."))
+    .configs(Integration)
+    .settings(commonSettings: _*)
+    .settings(libraryDependencies ++= dep)
+    .settings(inConfig(Integration)(Defaults.testSettings) : _*)
