+++ src/test/scala/io/iohk/ethereum/mpt/MerklePatriciaTrieSuite.scala
@@ -1,28 +1,27 @@
 package io.iohk.ethereum.mpt
 
 import java.io.File
+import java.nio.ByteBuffer
+import java.security.MessageDigest
 
-import akka.util.ByteString
 import io.iohk.ethereum.ObjectGenerators
 import io.iohk.ethereum.crypto.sha3
 import io.iohk.ethereum.mpt.MerklePatriciaTrie.defaultByteArraySerializable
-import io.iohk.ethereum.rlp.RLPImplicits._
-import io.iohk.ethereum.rlp.{encode => encodeRLP, decode => decodeRLP}
+import io.iohk.ethereum.rlp.{decode => decodeRLP, encode => encodeRLP}
 import io.iohk.iodb.LSMStore
 import org.scalacheck.{Arbitrary, Gen}
 import org.scalatest.FunSuite
 import org.scalatest.prop.PropertyChecks
 import org.spongycastle.util.encoders.Hex
-import scala.util.{Random, Try}
-import java.security.MessageDigest
-import java.nio.ByteBuffer
 
-class MerklePatriciaTreeSuite extends FunSuite
+import scala.util.{Random, Try}
+
+class MerklePatriciaTrieSuite extends FunSuite
   with PropertyChecks
   with ObjectGenerators {
   val hashFn = (input: Array[Byte]) => sha3(input)
 
-  val EmptyTrie = MerklePatriciaTrie[Array[Byte], Array[Byte]](HashMapDataSource(), hashFn)
+  val EmptyTrie = MerklePatriciaTrie[Array[Byte], Array[Byte]](EphemDataSource(), hashFn)
 
   implicit val intByteArraySerializable = new ByteArraySerializable[Int] {
     override def toBytes(input: Int): Array[Byte] = {
@@ -41,7 +40,7 @@ class MerklePatriciaTreeSuite extends FunSuite
   /* Random get, insert and delete tests */
   test("PatriciaTrie insert and get") {
     forAll(keyValueListGen()) { keyValueList: Seq[(Int, Int)] =>
-      val trie = keyValueList.foldLeft(MerklePatriciaTrie[Int, Int](HashMapDataSource(), hashFn)) {
+      val trie = keyValueList.foldLeft(MerklePatriciaTrie[Int, Int](EphemDataSource(), hashFn)) {
         case (recTrie, (key, value)) => recTrie.put(key, value)
       }
       keyValueList.foreach { case (key, value) =>
@@ -55,7 +54,7 @@ class MerklePatriciaTreeSuite extends FunSuite
   test("PatriciaTrie delete") {
     forAll(Gen.nonEmptyListOf(Arbitrary.arbitrary[Int])) { keyList: List[Int] =>
       val keyValueList = keyList.distinct.zipWithIndex
-      val trieAfterInsert = keyValueList.foldLeft(MerklePatriciaTrie[Int, Int](HashMapDataSource(), hashFn)) {
+      val trieAfterInsert = keyValueList.foldLeft(MerklePatriciaTrie[Int, Int](EphemDataSource(), hashFn)) {
         case (recTrie, (key, value)) => recTrie.put(key, value)
       }
       val (keyValueToDelete, keyValueLeft) = Random.shuffle(keyValueList).splitAt(Gen.choose(0, keyValueList.size).sample.get)
@@ -73,7 +72,7 @@ class MerklePatriciaTreeSuite extends FunSuite
         assert(obtained.isEmpty)
       }
 
-      val trieWithKeyValueLeft = keyValueLeft.foldLeft(MerklePatriciaTrie[Int, Int](HashMapDataSource(), hashFn)) {
+      val trieWithKeyValueLeft = keyValueLeft.foldLeft(MerklePatriciaTrie[Int, Int](EphemDataSource(), hashFn)) {
         case (recTrie, (key, value)) => recTrie.put(key, value)
       }
       assert(trieAfterDelete.getRootHash sameElements trieWithKeyValueLeft.getRootHash)
@@ -82,12 +81,12 @@ class MerklePatriciaTreeSuite extends FunSuite
 
   test("Trie insert should have the same root independently on the order its pairs are inserted") {
     forAll(keyValueListGen()) { keyValueList: Seq[(Int, Int)] =>
-      val trieAfterInsert = keyValueList.foldLeft(MerklePatriciaTrie[Int, Int](HashMapDataSource(), hashFn)) {
+      val trieAfterInsert = keyValueList.foldLeft(MerklePatriciaTrie[Int, Int](EphemDataSource(), hashFn)) {
         case (recTrie, (key, value)) => recTrie.put(key, value)
       }
       val keyValueListShuffle = Random.shuffle(keyValueList)
 
-      val trieAfterInsertShuffle = keyValueListShuffle.foldLeft(MerklePatriciaTrie[Int, Int](HashMapDataSource(), hashFn)) {
+      val trieAfterInsertShuffle = keyValueListShuffle.foldLeft(MerklePatriciaTrie[Int, Int](EphemDataSource(), hashFn)) {
         case (recTrie, (key, value)) => recTrie.put(key, value)
       }
 
@@ -97,13 +96,13 @@ class MerklePatriciaTreeSuite extends FunSuite
 
   /* MerklePatriciaTree API tests for particular cases */
   test("Remove key from an empty tree") {
-    val emptyTrie = MerklePatriciaTrie[Int, Int](HashMapDataSource(), hashFn)
+    val emptyTrie = MerklePatriciaTrie[Int, Int](EphemDataSource(), hashFn)
     val afterDeleteTrie = emptyTrie.remove(1)
     assert(afterDeleteTrie.getRootHash sameElements emptyTrie.getRootHash)
   }
 
   test("Remove a key that does not exist") {
-    val emptyTrie = MerklePatriciaTrie[Int, Int](HashMapDataSource(), hashFn)
+    val emptyTrie = MerklePatriciaTrie[Int, Int](EphemDataSource(), hashFn)
     val trieWithOneElement = emptyTrie.put(1, 5)
     val obtained = trieWithOneElement.get(1)
     assert(obtained.isDefined)
@@ -114,7 +113,7 @@ class MerklePatriciaTreeSuite extends FunSuite
   }
 
   test("Insert only one (key, value) pair to a trie and then deleted") {
-    val emptyTrie = MerklePatriciaTrie[Int, Int](HashMapDataSource(), hashFn)
+    val emptyTrie = MerklePatriciaTrie[Int, Int](EphemDataSource(), hashFn)
     val trieWithOneElement = emptyTrie.put(1, 5)
     val obtained = trieWithOneElement.get(1)
     assert(obtained.isDefined)
@@ -301,7 +300,7 @@ class MerklePatriciaTreeSuite extends FunSuite
     val key3: Array[Byte] = Hex.decode("123700")
     val key4: Array[Byte] = Hex.decode("123500")
     val trie = EmptyTrie.put(key1, key1).put(key2, key2).put(key3, key3).put(key4, key4)
-    val wrongSource = HashMapDataSource().update(
+    val wrongSource = EphemDataSource().update(
       rootHash = Array.emptyByteArray,
       toRemove = Seq(),
       toUpdate = Seq(trie.getRootHash -> trie.dataSource.get(trie.getRootHash).get)
@@ -366,84 +365,6 @@ class MerklePatriciaTreeSuite extends FunSuite
     val trieAfterDelete = trieWithOneElement.remove(1)
     val obtainedAfterDelete = trieAfterDelete.get(1)
     assert(obtainedAfterDelete.isEmpty)
-  }
-
-  ignore("IODB test - Insert of the first 5000 numbers hashed and then remove half of them"){
-    //create temporary dir
-    val dir = File.createTempFile("iodb", "iodb")
-    dir.delete()
-    dir.mkdir()
-
-    val dataSource = new IodbDataSource(new LSMStore(dir = dir, keySize = 32))
-    val emptyTrie = MerklePatriciaTrie[Array[Byte], Array[Byte]](dataSource, hashFn)
-
-    val keys = (0 to 100).map(intByteArraySerializable.toBytes)
-    val trie = Random.shuffle(keys).foldLeft(emptyTrie) { case (recTrie, key) => recTrie.put(md5(key), key) }
-
-    // We delete have of the (key-value) pairs we had inserted
-    val trieAfterDelete = Random.shuffle(keys.take(100/2)).foldLeft(trie) { case (recTrie, key) => recTrie.remove(md5(key)) }
-
-    // We delete keys with no effect so as to test that is the case (and for more code coverage)
-    val trieAfterDeleteNoEffect = keys.take(100/2).foldLeft(trieAfterDelete) { case (recTrie, key) => recTrie.remove(md5(key)) }
-    assert(Hex.toHexString(trieAfterDeleteNoEffect.getRootHash) == "b0bfbf4d2d6f3c9863c27f41a087208131f775edd9de2cb66242d1e0981aa94c")
-  }
-
-  ignore("IODB Test - PatriciaTrie insert and get") {
-    forAll(keyValueListGen()) { keyValueList: Seq[(Int, Int)] =>
-      //create temporary dir
-      val dir = File.createTempFile("iodb", "iodb")
-      dir.delete()
-      dir.mkdir()
-
-      val dataSource = new IodbDataSource(new LSMStore(dir = dir, keySize = 32))
-      val trie = keyValueList.foldLeft(MerklePatriciaTrie[Int, Int](dataSource, hashFn)) {
-        case (recTrie, (key, value)) => recTrie.put(key, value)
-      }
-      keyValueList.foreach { case (key, value) =>
-        val obtained = trie.get(key)
-        assert(obtained.isDefined)
-        assert(obtained.get == value)
-      }
-    }
-  }
-
-  ignore("IODB Test - PatriciaTrie delete") {
-    forAll(Gen.nonEmptyListOf(Arbitrary.arbitrary[Int])) { keyList: List[Int] =>
-      //create temporary dir
-      val dirWithDelete = File.createTempFile("iodb", "iodb1")
-      dirWithDelete.delete()
-      dirWithDelete.mkdir()
-      val dataSourceWithDelete = new IodbDataSource(new LSMStore(dir = dirWithDelete, keySize = 32))
-
-      val keyValueList = keyList.distinct.zipWithIndex
-      val trieAfterInsert = keyValueList.foldLeft(MerklePatriciaTrie[Int, Int](dataSourceWithDelete, hashFn)) {
-        case (recTrie, (key, value)) => recTrie.put(key, value)
-      }
-      val (keyValueToDelete, keyValueLeft) = Random.shuffle(keyValueList).splitAt(Gen.choose(0, keyValueList.size).sample.get)
-      val trieAfterDelete = keyValueToDelete.foldLeft(trieAfterInsert) {
-        case (recTrie, (key, value)) => recTrie.remove(key)
-      }
-
-      keyValueLeft.foreach { case (key, value) =>
-        val obtained = trieAfterDelete.get(key)
-        assert(obtained.isDefined)
-        assert(obtained.get == value)
-      }
-      keyValueToDelete.foreach { case (key, value) =>
-        val obtained = trieAfterDelete.get(key)
-        assert(obtained.isEmpty)
-      }
-
-      val dirOnlyInsert = File.createTempFile("iodb", "iodb2")
-      dirOnlyInsert.delete()
-      dirOnlyInsert.mkdir()
-      val dataSourceOnlyInsert = new IodbDataSource(new LSMStore(dir = dirOnlyInsert, keySize = 32))
-
-      val trieWithKeyValueLeft = keyValueLeft.foldLeft(MerklePatriciaTrie[Int, Int](dataSourceOnlyInsert, hashFn)) {
-        case (recTrie, (key, value)) => recTrie.put(key, value)
-      }
-      assert(trieAfterDelete.getRootHash sameElements trieWithKeyValueLeft.getRootHash)
-    }
   }
 
   /* EthereumJ tests */
@@ -512,108 +433,4 @@ class MerklePatriciaTreeSuite extends FunSuite
     val storage = EmptyTrie.put(key1, key1).put(key2, key2).put(key3, key3)
     assert(Hex.toHexString(storage.getRootHash) == "a3d0686205c7ed10a85c3bce4118d5d559bcda47ca39e4dd4f09719958a179f1")
   }
-
-  ignore("EthereumJ compatibility - Insert of the first 40000 numbers"){
-    val shuffledKeys = Random.shuffle(0 to 40000).map(intByteArraySerializable.toBytes)
-    val trie = shuffledKeys.foldLeft(EmptyTrie) { case (recTrie, key) => recTrie.put(key, key) }
-    assert(Hex.toHexString(trie.getRootHash) == "3f8b75707975e5c16588fa1ba3e69f8da39f4e7bf3ca28b029c7dcb589923463")
-  }
-
-  ignore("EthereumJ compatibility - Insert of the first 20000 numbers hashed"){
-    val shuffledKeys = Random.shuffle(0 to 20000).map(intByteArraySerializable.toBytes)
-    val trie = shuffledKeys.foldLeft(EmptyTrie) { case (recTrie, key) => recTrie.put(md5(key), key) }
-
-    // We insert keys that should have no effect so as to test that is the case (and for more code coverage)
-    val trieAfterInsertNoEffect = shuffledKeys.take(20000/2).foldLeft(trie) { case (recTrie, key) => recTrie.put(md5(key), key) }
-    assert(Hex.toHexString(trieAfterInsertNoEffect.getRootHash) == "a522b23a640c5fdb726e3f9644863e8913fe86339909fe881957efa0c23cebaa")
-  }
-
-  ignore("EthereumJ compatibility - Insert of the first 20000 numbers hashed and then remove half of them"){
-    val keys = (0 to 20000).map(intByteArraySerializable.toBytes)
-    val trie = Random.shuffle(keys).foldLeft(EmptyTrie) { case (recTrie, key) => recTrie.put(md5(key), key) }
-
-    // We delete have of the (key-value) pairs we had inserted
-    val trieAfterDelete = Random.shuffle(keys.take(20000/2)).foldLeft(trie) { case (recTrie, key) => recTrie.remove(md5(key)) }
-
-    // We delete keys with no effect so as to test that is the case (and for more code coverage)
-    val trieAfterDeleteNoEffect = keys.take(20000/2).foldLeft(trieAfterDelete) { case (recTrie, key) => recTrie.remove(md5(key)) }
-    assert(Hex.toHexString(trieAfterDeleteNoEffect.getRootHash) == "a693b82dcc5a9e581e9bf9aa7af3aed31fe3eb61f97fd733ce44c9f9df2d7f45")
-  }
-
-  ignore("EthereumJ compatibility - Insert of the first 20000 numbers hashed (with some sliced)"){
-    val keys = (0 to 20000).map(intByteArraySerializable.toBytes)
-
-    // We slice some of the keys so that me test more code coverage (if not we only test keys with the same length)
-    val slicedKeys = keys.zipWithIndex.map{case (key, index) =>
-      val hashedKey = md5(key)
-      if(index%2==0) hashedKey.take(hashedKey.length/2) else hashedKey
-    }
-    val keyValuePairs = slicedKeys.zip(keys)
-
-    val trie = Random.shuffle(keyValuePairs).foldLeft(EmptyTrie) { case (recTrie, (key, value)) => recTrie.put(key, value) }
-    assert(Hex.toHexString(trie.getRootHash) == "46cde8656f3be6ce93ba9dcb1017548f44c65d1ea659ac827fac8c9ac77cf6b3")
-  }
-
-  ignore("EthereumJ compatibility - Insert of the first 20000 numbers hashed (with some sliced) and then remove half of them") {
-    val keys = (0 to 20000).map(intByteArraySerializable.toBytes)
-
-    // We slice some of the keys so that me test more code coverage (if not we only test keys with the same length)
-    val slicedKeys = keys.zipWithIndex.map { case (key, index) =>
-      val hashedKey = md5(key)
-      if (index % 2 == 0) hashedKey.take(hashedKey.length / 2) else hashedKey }
-    val keyValuePairs = slicedKeys.zip(keys)
-
-    val trie = Random.shuffle(keyValuePairs).foldLeft(EmptyTrie) { case (recTrie, (key, value)) => recTrie.put(key, value) }
-
-    assert(Hex.toHexString(trie.getRootHash) == "46cde8656f3be6ce93ba9dcb1017548f44c65d1ea659ac827fac8c9ac77cf6b3")
-
-    // We delete have of the (key-value) pairs we had inserted
-    val trieAfterDelete = Random.shuffle(keyValuePairs.take(20000 / 2)).foldLeft(trie) { case (recTrie, (key, _)) => recTrie.remove(key) }
-
-    assert(Hex.toHexString(trieAfterDelete.getRootHash) == "ae7b65dddd3ac0428082160cf3ceff0276cf6e6deaa23b42c4c156b50a459822")
-  }
-
-  /* Performance test */
-  ignore("Performance test (From: https://github.com/ethereum/wiki/wiki/Benchmarks)"){
-    val debug = false
-    val Rounds = 1000
-    val Symmetric = true
-
-    val start: Long = System.currentTimeMillis
-    val emptyTrie = MerklePatriciaTrie[Array[Byte], Array[Byte]](HashMapDataSource(), hashFn)
-    var seed: Array[Byte] = Array.fill(32)(0.toByte)
-
-    val trieResult = (0 until Rounds).foldLeft(emptyTrie){ case (recTrie, i) =>
-      seed = hashFn(seed)
-      if(!Symmetric) recTrie.put(seed, seed)
-      else{
-        val mykey = seed
-        seed = hashFn(seed)
-        val myval = if((seed(0) & 0xFF) % 2 == 1) Array[Byte](seed.last) else seed
-        recTrie.put(mykey, myval)
-      }
-    }
-    val rootHash = Hex.toHexString(trieResult.getRootHash)
-    if(debug){
-      println("Time taken(ms): " + (System.currentTimeMillis - start))
-      println("Root hash obtained: " + rootHash)
-    }
-    if(Symmetric) assert(rootHash.take(4) == "36f6" && rootHash.drop(rootHash.length-4) == "93a3")
-    else assert(rootHash.take(4) == "da8a" && rootHash.drop(rootHash.length-4) == "0ca4")
-  }
-}
-
-case class HashMapDataSource(storage: Map[ByteString, Array[Byte]]) extends DataSource {
-
-  override def get(key: Array[Byte]): Option[Array[Byte]] = storage.get(ByteString(key))
-
-  override def update(rootHash: Array[Byte], toRemove: Seq[Key], toUpdate: Seq[(Key, Value)]): DataSource = {
-    val afterRemoval = toRemove.foldLeft(storage)((storage, key) => storage - ByteString(key))
-    val afterUpdate = toUpdate.foldLeft(afterRemoval)((storage, toUpdate) => storage + (ByteString(toUpdate._1) -> toUpdate._2))
-    HashMapDataSource(afterUpdate)
-  }
-}
-
-object HashMapDataSource {
-  def apply(): HashMapDataSource = HashMapDataSource(Map())
 }
